Server/Server/flask/requirements.txt
amqp==5.3.1
attrs==25.1.0
Authlib==1.4.1
bcrypt==4.2.1
bidict==0.23.1
billiard==4.2.1
blinker==1.9.0
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
colorlog==6.9.0
cryptography==44.0.1
Deprecated==1.2.18
eralchemy==1.5.0
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Cors==5.0.0
Flask-HTTPAuth==4.8.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
flexcache==0.3
flexparser==0.4
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
humanize==4.12.1
idna==3.10
importlib_resources==6.5.2
influxdb-client==1.48.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
Pint==0.24.4
platformdirs==4.3.6
prompt_toolkit==3.0.51
psutil==7.0.0
psycogreen==1.0.2
psycopg==3.2.4
psycopg-binary==3.2.4
pycparser==2.22
pycryptodome==3.21.0
pyfiglet==1.0.2
Pygments==2.19.1
PyJWT==2.10.1
pymemcache==4.0.0
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-memcached==1.62
python-socketio==5.12.1
reactivex==4.0.4
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
sympy==1.13.3
typing_extensions==4.12.2
tzdata==2025.2
upgrade-requirements==1.7.0
urllib3==2.3.0
uWSGI==2.0.28
vine==5.1.0
wcwidth==0.2.13
websockets==15.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
zope.event==5.0
zope.interface==7.2
